requirements.txt
# Core dependencies
composio-core>=0.3.0
composio-toolset>=0.3.0

# Google Sheets integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
gspread>=5.10.0

# Notion integration
notion-client>=2.0.0

# Email functionality
smtplib2>=0.2.1
email-validator>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0
httpx>=0.24.0

# Web framework for webhook server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Logging and monitoring
loguru>=0.7.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code quality (dev dependencies)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (dev dependencies)
mkdocs>=1.5.0
mkdocs-material>=9.0.0